requirements.txt
#Django
Django

#MongoDB
mongoengine
pymongo
dnspython

# Web server (Render start command uses gunicorn)
gunicorn


# Gemini API
google-generativeai